создаст класс Smart; сделает popup наследником Smart
