+++ src/view/smart.js
@@ -1,0 +1,57 @@
+import AbstractView from './abstract.js';
+
+export default class Smart extends AbstractView {
+  constructor() {
+    super();
+    this._data = {};
+    this._commentData = {};
+  }
+
+  updateData(update) {
+    if (!update) {
+      return;
+    }
+
+    this._data = Object.assign(
+      {},
+      this._data,
+      update,
+    );
+
+    this.updateElement();
+  }
+
+  updateCommentData(update, justDataUpdating) {
+    if (!update) {
+      return;
+    }
+
+    this._commentData = Object.assign(
+      {},
+      this._commentData,
+      update,
+    );
+
+    if (justDataUpdating) {
+      return;
+    }
+
+    this.updateElement();
+  }
+
+  updateElement() {
+    const prevElement = this.getElement();
+    const parent = prevElement.parentElement;
+    this.removeElement();
+
+    const newElement = this.getElement();
+
+    parent.replaceChild(newElement, prevElement);
+
+    this.restoreHandlers();
+  }
+
+  restoreHandlers() {
+    throw new Error('Abstract method not implemented: resetHandlers'); }
+
+}
